Move batch and updates scenarios to separate packages
