+++ jbpm-benchmarks/kieserver-performance-tests/src/main/java/org/jbpm/test/performance/taskassigning/TaskAssigning.java
@@ -44,7 +44,7 @@ import org.kie.server.client.UserTaskServicesClient;
  *     created in this scenario.
  * </li>
  */
-abstract class TaskAssigning {
+public abstract class TaskAssigning {
     protected static final String CONTAINER_ID = "kieserver-assets";
     protected static final String PROCESS_ID = "test-jbpm-assignment.testTaskAssignment";
