+++ jbpm-benchmarks/kieserver-performance-tests/src/main/java/org/jbpm/test/performance/taskassigning/batch/TaskAssigningBatch.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.jbpm.test.performance.taskassigning;
+package org.jbpm.test.performance.taskassigning.batch;
 
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
@@ -26,6 +26,7 @@ import java.util.concurrent.TimeUnit;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
+import org.jbpm.test.performance.taskassigning.TaskAssigning;
 import org.kie.perf.SharedMetricRegistry;
 import org.kie.perf.scenario.IPerfTest;
 import org.kie.server.api.model.definition.TaskQueryFilterSpec;
