+++ jbpm-benchmarks/kieserver-performance-tests/src/main/java/org/jbpm/test/performance/taskassigning/batch/TaskAssigningBatch_300.java
@@ -14,17 +14,17 @@
  *  limitations under the License.
  */
 
-package org.jbpm.test.performance.taskassigning;
+package org.jbpm.test.performance.taskassigning.batch;
 
 import org.kie.perf.scenario.IPerfTest;
 
 /**
- * This scenario simulates a batch task assignment of 3.000 tasks. To achieve this, 1.000 processes,
+ * This scenario simulates a batch task assignment of 300 tasks. To achieve this, 100 processes,
  * that contain 3 user tasks each, are started. {@link TaskAssigningBatch} for more details.
  */
-public class TaskAssigningBatch_3000 extends TaskAssigningBatch implements IPerfTest {
+public class TaskAssigningBatch_300 extends TaskAssigningBatch implements IPerfTest {
 
-    public TaskAssigningBatch_3000() {
-        super(1_000);
+    public TaskAssigningBatch_300() {
+        super(100);
     }
 }
